+++ packages/analytics-util-types/src/index.js
@@ -2,6 +2,7 @@
 /*
   Constants for reuse
 */
+
 export const FUNCTION = 'function'
 export const STRING = 'string'
 export const UNDEFINED = 'undefined'
@@ -10,8 +11,31 @@ export const OBJECT = 'object'
 export const ARRAY = 'array'
 export const NUMBER = 'number'
 export const noOp = () => {}
+export const ANY = 'any'
+export const ALL = '*'
 
+/* ────────────────────
+Environment checks
+─────────────────────── */
+
+/** @type {Boolean} */
 export const isBrowser = typeof window !== UNDEFINED
+
+/** @type {Boolean} */
+export const isNode = typeof process !== UNDEFINED && process.versions != null && process.versions.node != null
+
+/** @type {Boolean} */
+export const isDeno = typeof Deno !== UNDEFINED && typeof Deno.core !== UNDEFINED;
+
+/** @type {Boolean} */
+export const isWebWorker = typeof self === OBJECT && self.constructor && self.constructor.name === 'DedicatedWorkerGlobalScope'
+
+/** @type {Boolean} */
+export const isJsDom = (isBrowser && window.name === 'nodejs') || (typeof navigator !== UNDEFINED && (navigator.userAgent.includes('Node.js') || navigator.userAgent.includes('jsdom')))
+
+/* ────────────────────
+Type checks
+─────────────────────── */
 
 /** 
  * Check if value is function.
@@ -20,6 +44,18 @@ export const isBrowser = typeof window !== UNDEFINED
  */
 export function isFunction(x) {
   return typeof x === FUNCTION
+}
+
+/** 
+ * Check if value is ES2015 `class`.
+ * @param x
+ * @return {x is Class}
+ */
+export function isClass(x) {
+  if (isFunction(x)) {
+    return /^class /.test(Function.prototype.toString.call(x))
+  }
+  return false
 }
 
 /** 
@@ -33,11 +69,28 @@ export function isString(x) {
 
 /** 
  * Check if value is number.
- * @param v
- * @return {v is Object}
+ * @param n
+ * @return {boolean}
+ * @example
+ * > isNumber(0)
+ * true
+ * > isNumber(1)
+ * true
+ * > isNumber(1.1)
+ * true
+ * > isNumber(0xff)
+ * true
+ * > isNumber(0644)
+ * true
+ * > isNumber(6.2e5)
+ * true
+ * > isNumber(NaN)
+ * false
+ * > isNumber(Infinity)
+ * true
  */
-export function isNumber(v) {
-  return typeof v === NUMBER && !isNaN(v)
+export function isNumber(n) {
+  return !isNaN(parseFloat(n))
 }
 
 /** 
@@ -89,6 +142,47 @@ export function isObject(obj) {
   return Object.getPrototypeOf(obj) === proto
 }
 
+/** 
+ * Is primative value
+ * @param x
+ * @return {boolean}
+ * @example
+   isPrimitive(true) =>  true
+   isPrimitive({}) => false
+   isPrimitive(0) =>  true
+   isPrimitive('1') =>  true
+   isPrimitive(1.1) =>  true
+   isPrimitive(NaN) =>  true
+   isPrimitive(Infinity) =>  true
+   isPrimitive(function() {}) => false
+   isPrimitive(Date), => false
+   isPrimitive(null) =>  true
+   isPrimitive(undefined) =>  true
+ */
+export function isPrimitive (x) {
+  if (x === null) return true
+  switch (typeof x) {
+    case STRING:
+    case NUMBER:
+    case 'symbol':
+    case UNDEFINED:
+    case BOOLEAN:
+      return true
+    default:
+      return false
+  }
+}
+
+/**
+ * Returns true if the input is a Promise.
+ * @param {*} input - The input to test
+ * @returns {boolean}
+ * @static
+ */
+export function isPromise(x) {
+  if (!x) return false
+  return !!(!isUndefined(Promise) && x instanceof Promise || x.then && isFunction(x.then))
+}
 
 /**
  * Check if value is regexp
@@ -136,6 +230,10 @@ export function isEmail(str) {
   return /^[a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/.test(str)
 }
 
+/* ────────────────────
+HTML Element checks
+─────────────────────── */
+
 /** 
  * @param obj
  * @return {obj is NodeList}
@@ -162,7 +260,6 @@ export function isForm(element) {
   return isElement(element) && element.nodeName === 'FORM'
 }
 
-
 /**
  * Check if DOM element is hidden
  * @param {HTMLElement|null|undefined} element
@@ -170,11 +267,11 @@ export function isForm(element) {
  * @return {boolean} 
  */
 export function isHidden(element, until) {
-  if (!element || getComputedStyle(element).visibility === "hidden") return true
+  if (!element || getComputedStyle(element).visibility === 'hidden') return true
 
   while (element) {
     if (until != null && element === until) return false
-    if (getComputedStyle(element).display === "none") return true
+    if (getComputedStyle(element).display === 'none') return true
     element = element.parentElement
   }